chess_bot/requirements.txt
selenium>=4.15.0
python-chess>=1.10.0
numpy>=1.24.0
webdriver-manager>=4.0.0
